+++ include/ignition/math/Angle.hh
@@ -58,9 +58,10 @@ namespace ignition
       /// \brief Constructor
       public: Angle();
 
-      /// \brief Copy Constructor
+      /// \brief Conversion Constructor
       /// \param[in] _radian Radians
-      public: Angle(double _radian);
+      // cppcheck-suppress noExplicitConstructor
+      public: Angle(const double _radian);
 
       /// \brief Copy constructor
       /// \param[in] _angle Angle to copy
@@ -87,11 +88,6 @@ namespace ignition
 
       /// \brief Normalize the angle in the range -Pi to Pi
       public: void Normalize();
-
-      /// \brief Assignment operator. Set this angle to the parameter
-      /// \param[in]  _angle Angle to copy
-      /// \return The new angle.
-      public: Angle &operator=(const Angle &_angle);
 
       /// \brief Return the angle's radian value
       /// \return double containing the angle's radian value