+++ include/ignition/math/Vector2.hh
@@ -44,7 +44,7 @@ namespace ignition
       /// \brief Constructor
       /// \param[in] _x value along x
       /// \param[in] _y value along y
-      public: explicit Vector2(const T &_x, const T &_y)
+      public: Vector2(const T &_x, const T &_y)
       {
         this->data[0] = _x;
         this->data[1] = _y;
@@ -227,7 +227,7 @@ namespace ignition
       public: friend inline Vector2<T> operator-(const T _s,
                                                  const Vector2<T> &_v)
       {
-        return Vector2<T>(_s - _v.X(), _s - _v.Y());
+        return {_s - _v.X(), _s - _v.Y()};
       }
 
       /// \brief Subtraction assignment operator