+++ tools/code_check.sh
@@ -72,7 +72,7 @@ echo "*:include/ignition/math/Vector4.hh:257" >> $SUPPRESS
 echo "*:include/ignition/math/Vector4.hh:258" >> $SUPPRESS	
 
 #cppcheck
-CPPCHECK_BASE="cppcheck -q --suppressions-list=$SUPPRESS"
+CPPCHECK_BASE="cppcheck -q --suppressions-list=$SUPPRESS --inline-suppr"
 if [ $CPPCHECK_LT_157 -eq 0 ]; then
   # use --language argument if 1.57 or greater (issue #907)
   CPPCHECK_BASE="$CPPCHECK_BASE --language=c++"